Add writePullSecret()
Factored some code out of experimentalUpdateLayeredConfig into a
function in daemon.go since eventually we'll want to write pull secret
on daemon startup. We don't handle layering there currently so for now
we'll continue to call the function in experimentalUpdateLayeredConfig

Also moved getPullSecret to daemon
